save commit

+++ src/main/java/com/softserve/paymentservice/converter/InvoiceToDto.java
@@ -12,7 +12,7 @@ import java.math.RoundingMode;
 public class InvoiceToDto implements Converter<Invoice, InvoiceDto> {
     @Override
     public InvoiceDto convert(Invoice invoice) {
-        BigDecimal tripCost = BigDecimal.valueOf(invoice.getAmount()).divide(BigDecimal.valueOf(100), RoundingMode.HALF_UP); //todo test
+        BigDecimal tripCost =invoice.getAmount().divide(BigDecimal.valueOf(100), RoundingMode.HALF_UP); //todo test
         InvoiceDto invoiceDto = new InvoiceDto();
         invoiceDto.setUserId(invoice.getUserId());
         invoiceDto.setPaymentDate(invoice.getDateCreated());
